refactor(tools): extract use-tool-page-context composable to eliminate setup duplication
Identical ~15-line setup block (i18n, toast, session, entitlement, formatBrl)
was duplicated verbatim in hora-extra/page.vue and thirteenth-salary/page.vue,
pushing Sonar code duplication above the 3% threshold.

Extracted into app/features/tools/composables/use-tool-page-context.ts and
replaced both inline blocks with a single composable call.
