+++ app/features/tools/thirteenth-salary/page.vue
@@ -14,16 +14,12 @@ import {
   NTag,
   NThing,
   NTooltip,
-  useMessage,
 } from "naive-ui";
 import { Info } from "lucide-vue-next";
-import { useRouter } from "#app";
 
 import { captureException } from "~/core/observability";
-import { useApiError } from "~/composables/useApiError";
 import { useCalculatorFormState } from "~/features/tools/composables/use-calculator-form-state";
-import { useSessionStore } from "~/stores/session";
-import { useEntitlementQuery } from "~/features/paywall/queries/use-entitlement-query";
+import { useToolPageContext } from "~/features/tools/composables/use-tool-page-context";
 import { useSaveSimulationMutation } from "~/features/simulations/queries/use-save-simulation-mutation";
 import { useCreateGoalMutation } from "~/features/goals/queries/use-create-goal-mutation";
 import { useCreateReceivableMutation } from "~/features/receivables/queries/use-create-receivable-mutation";
@@ -47,21 +43,8 @@ import UiSurfaceCard from "~/components/ui/UiSurfaceCard/UiSurfaceCard.vue";
 
 defineOptions({ name: "ThirteenthSalaryPage" });
 
-const { t, n } = useI18n();
-const toast = useMessage();
-const { getErrorMessage } = useApiError();
-const router = useRouter();
-const sessionStore = useSessionStore();
-
-// ─── Session & access ─────────────────────────────────────────────────────────
-
-const isAuthenticated = computed<boolean>(() => sessionStore.isAuthenticated);
-
-const premiumAccessQuery = useEntitlementQuery("advanced_simulations");
-
-const hasPremiumAccess = computed<boolean>(
-  () => premiumAccessQuery.data.value === true,
-);
+const { t, toast, getErrorMessage, router, isAuthenticated, hasPremiumAccess, formatBrl } =
+  useToolPageContext();
 
 // ─── Calculator form state ────────────────────────────────────────────────────
 
@@ -89,18 +72,6 @@ const createReceivableMutation = useCreateReceivableMutation();
 const showGoalModal = ref(false);
 const goalForm = reactive({ name: "", targetDate: null as number | null });
 const showBudgetModal = ref(false);
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-/**
- * Formats a numeric value as Brazilian Real currency string.
- *
- * @param value Number to format.
- * @returns Formatted BRL string.
- */
-function formatBrl(value: number): string {
-  return n(value, "currency");
-}
 
 /**
  * Returns the ISO date string for the current year's target month and day.
